Created evaluate_peak.py
This module has 2 functions. It is able to locate and return the peak energy withing a specified energy range. It can also integrate the peak withing the energy range. A key element of this integration is reducing the spectrum data by subtracting a baseline reading which is achieved by creating a line across the lowest points of the peak and subtracting everything from below it. The SciPy Simps function is used to integrate.