Renamed Fluorescence_yield module
The fluorescence yield module takes as inputs the name of the element and the shell it peaks at as strings.
Output is the fluorescence yield associated with this.

This is a trivial thing to do in a module, but it makes the main.py look a lot neater to have it this way.