Ratio of elements improvement
I have improved the way the program calculates and prints the ratios of the elements of interest. I have made sure that all elements must be input in a particular order and the last 2 elements are the ones of interest. These last 2 elements have the ratios calculated and printed.

In a FAPbBr perovskite for example this would calculate the ratio of Pb:Br.
In a CsPbBrCl perovskite, this would be the ratio Br:Cl.