Removed redundant try, except statement. Improved ratio calculation, reformatted residual plot.
I have removed a redundant try-except statement from the Fluorescence section.

The method of calculating the ratio of Pb:halide has been changed and addded to the xrf_module.py. The ratio calculation here has been replaced by a call to that function.

The residual plot has been rid of the line of best fit. There is now a y=0 line for easy reference to zero.